perf: isolate kanban delete-drop-zone so drag start/end does not re-render the board
